fixed animation plot, fixed night plan plot, added wiyn telescope, added animation boundaries as variables, created workflow of input file, added tutorial and modified readme
