+++ README.md
@@ -1,8 +1,8 @@
 # ttpsolver
-Astronomical observation planning tool using the Traveling Telescope Problem (TTP) formulation. 
+Astronomical observation planning tool that solves the Traveling Telescope Problem (TTP), a time dependent version of the Traveling Salesman Problem.
 
 ## Purpose
-`ttpsolver` helps you plan observations for over 100 targets within a specified time frame. It simplifies the process by allowing you to input your observatory details, observing time constraints, and a target table. The tool plans observations to maximize successful exposures and optimizes the path between targets.
+`ttpsolver` helps you make an observing plan for your targets within a specified time frame. It simplifies the process by allowing you to input your observatory details, observing time constraints, and a target table. The tool plans observations to maximize the number of completed exposures and optimizes the slew path between targets.
 
 ## Installation
 To install `ttpsolver`, run:
@@ -32,9 +32,15 @@ conda install -c gurobi gurobi
 
 ## Algorithm
 
-`ttpsolver` employs a Mixed-Integer Linear Programming formulation to optimize scheduling in a time-dependent manner, similar to the Traveling Salesman Problem. It determines both the order of targets and the timing of observations to respect celestial object rise and set times. The algorithm is detailed in the Astronomical Journal [here](https://iopscience.iop.org/article/10.3847/1538-3881/ad0dfb). Please cite this work when using ttpsolver for planning scientific observations.
+`ttpsolver` employs a Mixed-Integer Linear Programming formulation to optimize scheduling in a time-dependent manner, similar to the Traveling Salesman Problem. It determines both the order of targets and the timing of observations to respect celestial object rise and set times. The algorithm is detailed in the Astronomical Journal [here](https://iopscience.iop.org/article/10.3847/1538-3881/ad0dfb). Please cite this work when using `ttpsolver` for planning scientific observations.
 
 ## Tutorials
-You can find a simple tutorial for using the package in the `/tutorial` directory.
+You can find a tutorial for using the package in the `/tutorial` directory. In the tutorial, we attempt to schedule 85 targets (the survey list for the TESS-Keck Survey, see [here](https://ui.adsabs.harvard.edu/abs/2022AJ....163..297C/abstract), as this is a large list of stars spread out over RA hours). Not all stars will be observable on the specified night but the `ttpsolver` handles this easily.
 
+To run `ttpsolver`, we require an inputs file where the user specifies information about the logistics of the night (start and stop time to observations and the telescope to be used), as well as a file containing the table of targets and their observational strategy. Each target requires RA/Dec in hour angle/degrees, and an exposure time in seconds. Next each target's strategy is described by the number of exposures to take at each visit, the number of visits to make to the target and the minimum separation in time (hours) between visits to the same target. Lastly, each target requires a priority ranking from 1 to 10, with 10 being the highest priority.
 
+Feel free to play with this tutorial in changing the exposure times, the observational strategies, the priorities, and/or the time window to compute the optimal path (both in terms of calendar date and hour window within the night).
+
+`ttpsolver` returns a few reports and plots. First, it outputs a .txt file where the optimal slew path is laid out as each each target is listed next to the starting time of the exposure. Next, it outputs a .txt file with statistics about the solution. Then it generates 3 plots: a Night Plan (both as an interactive .html and as a static png), a 2D slew path representation, and an animation portraying the slew path over the course of the night.
+
+We built `ttpsolver` with two telescopes pre-made: Keck1 and WIYN. However, we designed the telescope class to be easily replicated for any other custom observatory. See instructions within the telescope.py file.
