+++ tutorial/inputs.txt
@@ -1,0 +1,5 @@
+StarListFile : ../tutorial/TKStargets.csv
+StartNight : 2024-02-11T07:00:00
+EndNight : 2024-02-11T14:00:00
+Telescope : Keck1
+SavePath : ../tutorial/outputs/
